move: 서버 컴포넌트와 클라이언트 컴포넌트 분리 (#193)
