+++ src/app/(mypage)/account/_components/_client/account-profile-image.tsx
