+++ src/app/(mypage)/account/page.tsx
@@ -3,9 +3,9 @@ import { Suspense } from 'react';
 import Loading from '@/app/loading';
 import { fetchGetCurrentUser } from '@/lib/apis/auth/auth-server.api';
 
-import PersonalInfo from './_components/account-personal-info';
-import ProfileInfo from './_components/account-profile-info';
-import SecurityInfo from './_components/account-security-info';
+import PersonalInfo from './_components/_client/account-personal-info';
+import ProfileInfo from './_components/_client/account-profile-info';
+import SecurityInfo from './_components/_client/account-security-info';
 
 export const metadata = {
   title: '마이페이지 - 회원정보 수정',
