+++ src/app/(mypage)/bookmarks/[category]/page.tsx
@@ -1,5 +1,5 @@
 import { CategoryParamType } from '@/types/category.type';
-import BookmarksList from '../_components/bookmarks-list';
+import BookmarksList from '../_components/_client/bookmarks-list';
 
 export const metadata = {
   title: '마이페이지 - 내가 북마크한 장소',
